GenAi_Notebooks/5_deployment/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
gunicorn>=20.1.0
boto3>=1.26.0
google-cloud-storage>=2.5.0
azure-storage-blob>=12.14.0
kubernetes>=24.2.0
python-kubernetes>=24.2.0
